pyproject.toml
[project]
name = "oracle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "commitizen>=4.4.1",
    "fastapi>=0.115.11",
    "jupyter>=1.1.1",
    "langchain>=0.3.20",
    "langchain-anthropic>=0.3.9",
    "langchain-community>=0.3.19",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "langsmith>=0.3.13",
    "ollama>=0.4.7",
    "pre-commit>=4.1.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
